refactor: ItemMenuManager 코드 분리후 정리
ItemMenuManager에서 BuyCatMenu에 대한 코드를 작성했었는데 BuyCatMenuManger 스크립트로 분리함. 추가로 UI도 최신화함
