+++ Cat_Merge/Assets/2.Prefabs/BuyCatMenu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c9eff685ca0e72a4283eee11e5db9243
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
