+++ Cat_Merge/Assets/2.Prefabs/ItemMenu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b0ab1f814d3e9ab4592bfda7d95be1ac
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
