+++ Cat_Merge/Assets/2.Prefabs/ItemMenu/둘기 최대치 증가.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f2919b5a5d8e3ee458460fc702a3b940
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
